requirements.txt
bidict
click
cycler
joblib
kiwisolver
kneed
loguru
matplotlib
numpy
pandas
pyparsing
python-dateutil
pytz
scikit-learn
scipy
seaborn
six
pylru
scipy
